Moved CRcvBuffer2 to rcvbuffer.h
